+++ src/main/java/org/zalando/nakadi/validation/schema/diff/ReferenceSchemaDiff.java
@@ -1,18 +1,14 @@
 package org.zalando.nakadi.validation.schema.diff;
 
 import org.everit.json.schema.ReferenceSchema;
-import org.zalando.nakadi.domain.SchemaChange;
-
-import java.util.List;
-import java.util.Stack;
 
 public class ReferenceSchemaDiff {
-    static void recursiveCheck(final ReferenceSchema referenceSchemaOriginal,
-                                       final ReferenceSchema referenceSchemaUpdate, final Stack<String> jsonPath,
-                                       final List<SchemaChange> changes) {
-        jsonPath.push("$ref");
-        SchemaDiff.recursiveCheck(referenceSchemaOriginal.getReferredSchema(),
-                referenceSchemaUpdate.getReferredSchema(), jsonPath, changes);
-        jsonPath.pop();
+    static void recursiveCheck(
+            final ReferenceSchema referenceSchemaOriginal, final ReferenceSchema referenceSchemaUpdate,
+            final SchemaDiffState state) {
+        state.runOnPath("$ref", () -> {
+            SchemaDiff.recursiveCheck(referenceSchemaOriginal.getReferredSchema(),
+                    referenceSchemaUpdate.getReferredSchema(), state);
+        });
     }
 }